fix(footer): don't put GitHub logo on a separate line
We had a grid with `grid-cols-1 sm:grid-cols-2`. A paragraph and the
logo were present in separate grid cells. This meant that the GitHub
logo was on a separate line to the text at small widths.

Fix this by splitting a grid: we want the appearance selector to have
this 2 -> 1 behaviour, but the GitHub logo should always be on the same
line as the link to my website.