feat: add API to generate 10PRINT SVGs
Add an API `/api/svg` to generate 10PRINT SVGs. This should allow them
to be used from other webpages.

Factor the generation out to a library which generates DOM nodes either
directly using the browser's API or via `linkedom` when running on the
server.

Add tests for this using `vitest`, and run these in Actions CI.