feat: add static JS file to call `/api/svg` endpoint
We added this endpoint recently. Here is a JS file served as
`/static/background.js` which calls the endpoint and sets the SVG as the
`<body>` element's background. A debounced event listener is also put in
place to resize on window size change, similar to what we have in the
React frontend.

The JS file only supports being imported and executed as a main module
at the minute. Next will be a change to let it be called, so that it can
take parameters and be applied to arbitrary elements.