feat: add static JS files to call `/svg` endpoint
We added this endpoint recently. Here are a couple of JS files served as
`/background{,-body}.js` which call the endpoint and sets the SVG as the
`<body>` (or an arbitrary) element's background. A debounced event
listener is also optionally put in place to resize on window size
change, similar to what we have in the React frontend.

Rename the existing `/api/svg` endpoint to `/svg`. It's going to be the
only one, so make it simple.