feat: support saving the seed under a different name
Our demo application has a body and an element using `10PRINT` SVGs.
Since we didn't support customising the seed key, after the initial load
(where no key is stored in local storage), they would use the same.

Support supplying a key, so that elements can set it. The body doesn't
accept parameters right now so it can't do that. But this is enough to
have distinct seeds.