chore: don't list `.{m,}js` files in the `tsconfig.json` file
This results in errors from `tsc` that it wants to overwrite the same
files. I think because it outputs `.js` and we don't specify an output
directory so it'll output in the same dir, overwriting the file.

Typechecking the Javascript can also be achieved by `// @ts-check`
comments, so do that instead.