chore(renovate): group non-major updates, replace some config with presets
The current config creates a separate PR for each update. Since we're
using npm packages, this is a _lot_. Run Renovate once a day, and
configure it to group all non-major updates together, so that we get far
fewer.

While we're here, some of the config can be removed and replaced with
presets, just to simplify things a bit.