feat: randomise colours button
Add a function to generate random colours that go well together and hook
it up in the UI. We're using culori for this. The config interface
remains the same, but internally we parse to a culori object.

The sheet (sidebar) is slightly reworked to be a bit more consistent and
allow for the new button in a clean way.